Add Quotes Notes

+++ src/components/comments/comment.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable no-alert */
-import { useRouter } from 'next/router'
 import React, { useEffect, useState } from 'react'
-import { HiOutlineDotsVertical } from 'react-icons/hi'
 import styles from 'sass/components/comments/comment.module.scss'
+import SpeechBubble from 'widgets/speech-bubble'
 import FormProvider from 'providers/form'
+import usePageQuery from 'lib/hooks/page-query'
 import Input from './input'
 import { useComments } from './inner'
 
@@ -16,22 +16,18 @@ export interface commentData {
   createdAt: string
 }
 
-export default function Comment({
-  data,
-}: {
-  data: commentData
-}) {
-  const router = useRouter()
-  const {
-    id, name, content,
-  } = data
+export default function Comment({ data }: { data: commentData }) {
+  const { category, current } = usePageQuery()
+  const { id, name, content } = data
   const [state, setState] = useState<commentState>()
-  const [extended, setExtended] = useState(false)
   const [password, setPassword] = useState(null)
   const { refresh } = useComments()
 
   const onEdit = async () => {
     const password = prompt('비밀번호를 입력하세요.')
+    if (!password) {
+      return
+    }
     const res = await fetch('/api/comments/validate', {
       body: JSON.stringify({
         id,
@@ -46,25 +42,31 @@ export default function Comment({
       setPassword(password)
       setState('edit')
     } else {
-      const { message } = await res.json()
-      alert(message)
+      const { error } = await res.json()
+      alert(error)
     }
   }
 
   const onDelete = async () => {
     const password = prompt('정말로 지우시겠습니까? 비밀번호를 입력하세요.')
-    const formData = new FormData()
-    formData.append('id', id)
-    formData.append('password', password)
-    const res = await fetch(`/api/comments${router.asPath}`, {
-      body: formData,
+    if (!password) {
+      return
+    }
+    const res = await fetch(`/api/comments/${category}/${current}`, {
+      body: JSON.stringify({
+        id,
+        password,
+      }),
+      headers: {
+        'Content-Type': 'application/json',
+      },
       method: 'DELETE',
     })
-    if (res.status === 200) {
+    if (res.ok) {
       refresh()
     } else {
-      const { message } = await res.json()
-      alert(message)
+      const { error } = await res.json()
+      alert(error)
     }
   }
 
@@ -88,35 +90,13 @@ export default function Comment({
       </FormProvider>
     )
   }
-
   return (
-    <div className={styles.container}>
-      <header className={styles.header}>
-        <span className={styles.name}>{name}</span>
-        <div className={styles.arrow} />
-      </header>
-      <div className={styles.body}>
-        <p className={styles.content}>{content}</p>
-        <div className={styles.sideMenu}>
-          <button
-            type="button"
-            className={styles.menuButton}
-            onClick={() => setExtended(!extended)}
-          >
-            <HiOutlineDotsVertical size={16} />
-          </button>
-          {extended && (
-            <div className={styles.menu}>
-              <div className={styles.buttonWrapper}>
-                <button type="button" onClick={onEdit}>수정</button>
-              </div>
-              <div className={styles.buttonWrapper}>
-                <button type="button" onClick={onDelete}>삭제</button>
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
+    <SpeechBubble
+      head={<span className={styles.name}>{name}</span>}
+      word={content}
+      onEdit={onEdit}
+      onDelete={onDelete}
+      needsAuth={false}
+    />
   )
 }
